+++ docs/changelog.md
@@ -11,6 +11,7 @@ Each revision is versioned by the date of the revision.
 ## 2026-01-16
 
 - Added terraform modules for the HAProxy DDoS Protection configurator charm.
+- Added path rewrite support for gRPC backends.
 
 ## 2026-01-15
 
@@ -19,6 +20,10 @@ Each revision is versioned by the date of the revision.
 ## 2026-01-15
 
 - Added missing settings from haproxy-route-tcp relation template.
+
+## 2026-01-14
+
+- Added provider support for the `ddos_protection` interface on the HAProxy charm.
 
 ## 2026-01-09